Move from Serilog.Expressions to Seq.Syntax
